Add script for training teacher.
